move date funcitons into lib
