+++ ui/src/lib.ts
@@ -11,4 +11,41 @@ const formatDate: Function = (date: Date): string => {
 
 const getDateTimeString = (date: string, time: string) => `${date}T${time}`;
 
-export { formatDate, getDateTimeString, padNumberWith0Zero };
+const modalDateFormat = (selectedEventDate: Date) =>
+  `${selectedEventDate.toLocaleString("default", {
+    weekday: "long",
+  })}, ${selectedEventDate.toLocaleString("default", {
+    month: "long",
+  })} ${selectedEventDate.getDate()}
+${selectedEventDate.toLocaleTimeString("default", {
+  hour: "2-digit",
+  minute: "2-digit",
+})}
+`;
+
+const allDayDateFormat = (selectedEventDate: Date) =>
+  `${selectedEventDate.toLocaleString("default", {
+    weekday: "long",
+  })}, ${selectedEventDate.toLocaleString("default", {
+    month: "long",
+  })} ${selectedEventDate.getDate()}
+`;
+
+const getModalDate = (allDay: boolean, dateTime: string) => {
+  const date = new Date(dateTime);
+  return allDay ? allDayDateFormat(date) : modalDateFormat(new Date(date));
+};
+
+const formatTime = (utcString: string) =>
+  `${padNumberWith0Zero(new Date(utcString).getHours())}:${padNumberWith0Zero(
+    new Date(utcString).getMinutes(),
+  )}`;
+
+export {
+  formatDate,
+  getDateTimeString,
+  padNumberWith0Zero,
+  modalDateFormat,
+  getModalDate,
+  formatTime,
+};
